==> code/cross-task-simil_jocn2021/1_est_simil_jocn_2trpk.R <==

source(here::here("code", "_packages.R"))
source(here("code", "read-behav.R"))
source(here("code", "_vars.R"))
source(here("code", "_atlases.R"))
source(here("code", "_settings.R"))
source(here("code", "_funs.R"))


n.vert <- 20484

glminfo <- data.frame(
  task = c("Axcpt", "Cuedts", "Stern", "Stroop"),
  name.glm = c(
    "baseline_Cues_EVENTS_censored",
    "baseline_CongruencyIncentive_EVENTS_censored",
    "baseline_ListLength_EVENTS_censored",
    "baseline_Congruency_EVENTS_censored"
  ),
  stringsAsFactors = FALSE
)
glminfo <- as.data.table(glminfo)
hi = list(
  Axcpt = "BX",
  Cuedts = c("InConInc", "InConNoInc"),
  Stern = "LL5RN",
  Stroop = c("biasInCon", "PC50InCon")
)
lo = list(
  Axcpt = "BY",
  Cuedts = c("ConInc", "ConNoInc"),
  Stern = "LL5NN",
  Stroop = c("biasCon", "PC50Con")
)

n.iter <- length(subjs)
pb <- progress_bar$new(
  format = " running [:bar] :percent eta: :eta (elapsed: :elapsed)",
  total = n.iter, clear = FALSE, width = 120
)


target.trs.2prk <- list(
  Axcpt = 4,
  Cuedts = 4,
  Stern = 6,
  Stroop = 2
)

# subjs <- subjs[!subjs %in% c("DMCC5820265")]
# subjs[subjs == "DMCC5820265"] <- "DMCC5820265_old-delete when rerun"


## wrangle betas ----
(time.start <- Sys.time())


conditions <- combo_paste(tasks, c("hi", "lo"))

betas <- 
  array(
    NA,
    dim = c(n.vert, length(conditions), length(subjs)),
    dimnames = list(vertex = NULL, condition = conditions, subj = subjs)
  )
for (glm.i in seq_len(nrow(glminfo))) {
  # glm.i = 2
  
  name.glm.i <- glminfo$name.glm[glm.i]
  name.task.i <- glminfo$task[glm.i]
  
  ## read betas:
  betas.i <- readRDS(
    here(
      "out", "glms",
      paste0("betas_dmcc_2trpk_", name.task.i, "_", name.glm.i,  ".RDS")
    )
  )
  ## remove subj with missing data, get target TRs:
  betas.i <- betas.i[, , target.trs.2prk[[glm.i]], ]
  betas.i <- aperm(betas.i, c("vertex", "subj", "reg"))
  
  betas.i.hi <- rowMeans(betas.i[, , hi[[glm.i]], drop = FALSE], dim = 2) ## ave across TR, run, relevant regs
  betas.i.lo <- rowMeans(betas.i[, , lo[[glm.i]], drop = FALSE], dim = 2)
  betas.ii <- abind(betas.i.hi, betas.i.lo, along = 0)
  
  # ## average across target TRs:
  # betas.i <- abind(
  #   apply(betas.i[, hi[[glm.i]], target.trs[[glm.i]], , ], c("vertex", "subj", "run"), mean),
  #   apply(betas.i[, lo[[glm.i]], target.trs[[glm.i]], , ], c("vertex", "subj", "run"), mean),
  #   along = 0
  # )   ## condition, vertex, subj
  # 
  # betas.ii <- (betas.i[, , , "run1"] + betas.i[, , , "run2"]) / 2  ## average across run
  
  
  ## reshape to match dim(betas):
  
  betas.ii <- aperm(betas.ii, c(2, 1, 3))
  conditions.i <- combo_paste(name.task.i, c("hi", "lo"))  ## run varies slower
  betas[, conditions.i, subjs] <- betas.ii
  
  
}
rm(betas.i)
gc()






## est simil ----


simil <- 
  array(
    NA,
    dim = c(length(conditions), length(conditions), length(parcellation$key), length(subjs)),
    dimnames = list(.row = conditions, .col = conditions, parcel = parcellation$key, subj = subjs)
  )

univa <- 
  array(
    NA,
    dim = c(length(conditions), length(conditions), length(parcellation$key), length(subjs)),
    dimnames = list(.row = conditions, .col = conditions, parcel = parcellation$key, subj = subjs)
  )

unifo <- 
  array(
    NA,
    dim = c(length(conditions), length(conditions), length(parcellation$key), length(subjs)),
    dimnames = list(.row = conditions, .col = conditions, parcel = parcellation$key, subj = subjs)
  )


for (subj.i in seq_along(subjs))  {
  # subj.i = 1
  
  res <- enlist(parcellation$key)
  
  
  name.subj.i <- subjs[subj.i]
  
  # if (name.subj.i %in% c("DMCC5820265")) next
  
  betas.subj.i <- betas[, , subj.i]
  
  
  for (parcel.i in seq_along(parcellation$key)) {
    # parcel.i = 20
    
    ## mask:
    
    is.parcel <- schaefer10k == parcel.i
    betas.subj.parcel.i <- betas.subj.i[is.parcel, ]
    
    has.signal.all.conds <- !rowSums(abs(betas.subj.parcel.i) < .Machine$double.eps) > 0
    if (sum(has.signal.all.conds) <= length(conditions)) stop(paste0("bad parcel: ", subj.i, " ", parcel.i))
    
    B <- betas.subj.parcel.i[has.signal.all.conds, ]
    
    
    ## estimate similarity matrices:
    
    simil[, , parcel.i, subj.i] <- cor(B)
    
    univa[, , parcel.i, subj.i] <- tcrossprod(colMeans(B))
    unifo[, , parcel.i, subj.i] <- tcrossprod(colSums(B^2))
    
  }
  
  pb$tick()  ## progress bar
  
  
}



## save ----

if (!dir.exists(here("out", "cross-task-simil_jocn2021"))) dir.create(here("out", "cross-task-simil_jocn2021"))
saveRDS(simil, here("out", "cross-task-simil_jocn2021", paste0("corr-biased_unpre_jocn_2trpk.RDS")))
saveRDS(univa, here("out", "cross-task-simil_jocn2021", paste0("mean-biased_unpre_jocn_2trpk.RDS")))
saveRDS(unifo, here("out", "cross-task-simil_jocn2021", paste0("ssq-biased_unpre_jocn_2trpk.RDS")))



(time.run <- Sys.time() - time.start)
