==> code/cross-task-simil_jocn2021/3_rsa_jocn_2trpk.rmd <==
---
title: 'multitask pattern similarity: high versus low, 2trpk'
author: "michael freund"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    highlight: zenburn
    number_sections: true
    theme: spacelab
---


# intro


## Purpose

supplementary analysis for jocn 


```{r setup, include = FALSE}

source(here::here("code", "_packages.R"))
source(here("code", "read-behav.R"))
source(here("code", "_vars.R"))
source(here("code", "_atlases.R"))
source(here("code", "_settings.R"))
source(here("code", "_funs.R"))

library(ggridges)
library(colorspace)

# subjs <- subjs[!subjs %in% "DMCC5820265"]
conditions <- combo_paste(tasks, c("hi", "lo"))
n.dim <- length(conditions)
is.lower.tri <- lower.tri(diag(n.dim))
tris <- fread(here("out", "multitask", "rsa-model-tris_cross-run.csv")) %>% 
  filter(grepl("run1", .row), grepl("run1", .col))
X <- tris %>% select(where(is.numeric)) %>% as.matrix

rsarray <- readRDS(here("out", "cross-task-simil_jocn2021", paste0("corr-biased_unpre_jocn_2trpk.RDS")))


group.target.trs <- fread(here("out", "stats_univ_hilo_target_schaefer400-07.csv"))
dmcc55b.parcels <- group.target.trs %>% filter(estimate > 0, p.fdr < 0.05) %>% pull(parcel) %>% table
dmcc55b.parcels <- names(dmcc55b.parcels)[dmcc55b.parcels == 4]


stats.subjs <- fread(here("out", "cross-task-simil_jocn2021",  paste0("subjs_jocn.csv")))
stats.subjs.univa <- fread(here("out", "multitask",  paste0("subjs_jocn_mean.csv")))
stats.subjs.unifo <- fread(here("out", "multitask",  paste0("subjs_jocn_ssq.csv")))


theme_mat <- list(
  theme(axis.text.y = element_text(), axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))
  )


dmcc35 <- c(
  77, 78,  86,  87,  88, 90, 91,  99, 101, 103, 105, 110, 127, 130, 132, 139, 140, 141, 
  148, 172, 175, 185, 189, 290, 306, 311, 314, 337, 340, 346, 347, 349, 350, 353, 361
  )

null34 <- c(
  4,5,51,72,97,100,121,122,123,125,150,153,174,211,221,235,244,245,
  266,268,270,272,279,302,304,307,321,323,325,326,330,369,376,381
)


```



## Notes


**Data/sample/preproc**

  * UB55 sans DMCC52*
  * surface + fMRIprep
  * schaefer 400-07
  
**1st-Level GLM**

  * 2trpk, runwise GLMs
  
Target TRs:

  * Axcpt: `r target.trs$Axcpt`
  * Cuedts: `r target.trs$Cuedts`
  * Stern: `r target.trs$Stern`
  * Stroop: `r target.trs$Stroop`

GLM contrasts:

  * Axcpt: $\text{BY} - \text{BX}$
  * Cuedts: $(\text{InConSwitch} + \text{InConRepeat} - \text{ConSwitch} - \text{ConRepeat})/2$
  * Stern: $\text{LL5RN} - \text{LL5NN}$
  * Stroop: $(\text{PC50InCon} + \text{biasInCon} - \text{PC50Con} - \text{biasCon})/2$


**RSA**

Similarity measure:

  * Similarity measure: standard linear correlation
  * NOT cross-valdiated (e.g., not cv-Euclidean/Mahalanobis)
  * NOT prewhitened
  * betas averaged across run, within task, for RSA

models:

  * focused only on between-task correlations
  * atanh transformed


```{r fig.width = 16, echo = FALSE}

## build models

rsm.empty <- matrix(0, ncol = n.dim, nrow = n.dim, dimnames = list(conditions, conditions))
rsm.task <- rsm.empty
rsm.hi <- rsm.empty
rsm.lo <- rsm.empty
rsm.hilo <- rsm.empty

for (task.i in tasks) rsm.task[grepl(task.i, conditions), grepl(task.i, conditions)] <- 1
rsm.hi[grepl("hi", conditions), grepl("hi", conditions)] <- 1
rsm.lo[grepl("lo", conditions), grepl("lo", conditions)] <- 1
rsm.hilo[grepl("lo", conditions), grepl("hi", conditions)] <- 1
rsm.hilo[grepl("hi", conditions), grepl("lo", conditions)] <- 1

## plot


plot_grid(
  plot_grid(
    rsm.task %>% matplot + theme_mat + labs(title = "task (cells excluded)"),
    rsm.hi   %>% matplot + theme_mat + labs(title = "hihi"),
    rsm.lo   %>% matplot + theme_mat + labs(title = "lolo"),
    rsm.hilo   %>% matplot + theme_mat + labs(title = "hilo"),
    nrow = 1
  ),
  rel_heights = c(0.1, 1)
)


```



# quick look: all correlations

```{r fig.width = 10, fig.height = 7}



p.allcor.density <- stats.subjs %>%
  
  ggplot(aes(beta, y = term, fill = term)) +
  geom_density_ridges() +
  
  scale_fill_brewer(type = "qual", palette = 5) +
  
  theme(legend.position = "none") +
  labs(x = "cor (z)")


p.allcor.box <- stats.subjs %>%
  
  ggplot(aes(term, beta, fill = term)) +
  geom_hline(yintercept = 0) +
  geom_boxplot(width = 0.1, notch = TRUE) +
  
  scale_fill_brewer(type = "qual", palette = 5) +
  
  theme(legend.position = "none") +
  labs(y = "cor (z)") 



p.allcor.network <- stats.subjs %>%
  
  ggplot(aes(beta, y = term, fill = term)) +
  geom_density_ridges() +
  
  scale_fill_brewer(type = "qual", palette = 5) +
  
  facet_grid(
    rows = vars(get.network(parcel)), 
    labeller = labeller(run = c(bt = "between-run", wn = "within-run")),
    scales = "free_x"
    ) +
  
  theme(legend.position = "none") +
  labs(y = "cor (z)")


plot_grid(
  cowplot.title("Pattern correlations by components of RSA model"),
  plot_grid(plot_grid(p.allcor.density, p.allcor.box, ncol = 1), p.allcor.network, nrow = 1),
  ncol = 1, 
  rel_heights = c(0.1, 1)
)


```


# Parcel-wise stats


```{r}

contrasts.subjs <- stats.subjs %>%
  
  pivot_wider(names_from = "term", values_from = "beta") %>%
  
  mutate(
    hihi_lolo = hihi - lolo,
    hihi_hilo = hihi - hilo,
    lolo_hilo = lolo - hilo
    ) %>%
  
  pivot_longer(
    cols = c("hihi_hilo", "lolo_hilo", "hihi", "lolo", "hilo", "hihi_lolo"), 
    names_to = "contrast", values_to = "b"
    )



group.contrasts <- contrasts.subjs %>%
  
  group_by(contrast, parcel) %>%
  
  summarize(
    p = t.test(b)$p.value,
    statistic = t.test(b)$statistic,
    b = mean(b),
    .groups = "drop_last"
    ) %>%
  
  group_by(p.fdr = p.adjust(p, "fdr"))


```


* displayed are t statistics
* thresholded images are thesholded at and t > 0 and p < 0.05 FDR corrected across all 400 parcels

```{r, fig.width = 15, results = "asis"}

contrs <- unique(group.contrasts$contrast)


group.contrasts %<>%
  
  mutate(
    hemi = substr(parcel, 1, 1),
    num.roi = match(parcel, parcellation$key)
  ) %>%
  
  arrange(contrast, num.roi)



for (contrs.i in contrs) {
  
  cat("\n ")
  cat(sprintf("\n### %s ", contrs.i))
  cat("\n ")
  
  group.contrasts %>%

    filter(contrast == contrs.i) %>%

    build_overlay("statistic", template = schaefer) %>%
    plot_surface(underlay = hcp, hues = rev(viridis::inferno(500)))


  group.contrasts %>%

    filter(contrast == contrs.i) %>%
    mutate(statistic = ifelse(p.fdr < 0.05 & b > 0, statistic, 0)) %>%

    build_overlay("statistic", template = schaefer) %>%
    plot_surface(underlay = hcp, hues = rev(viridis::inferno(500)))

  
}


```


# DMCC 35 versus null 35


```{r}

contrasts.subjs.35 <-
  
  contrasts.subjs %>%
  
  filter(
    parcel %in% parcellation$key[c(dmcc35, null34)], 
    contrast %in% c("hihi_lolo", "hihi_hilo", "lolo_hilo")
    ) %>%
  mutate(
    parcelset = ifelse(parcel %in% parcellation$key[dmcc35], "dmcc35", "null34")
  ) %>%
  
  group_by(subj, contrast, parcelset) %>%
  summarize(b = mean(b))



contrasts.subjs.35 %>%
  
  group_by(contrast, parcelset) %>%
  summarize(
    p = t.test(b)$p.value,
    statistic = t.test(b)$statistic,
    b = mean(b),
    .groups = "drop_last"
    )

contrasts.subjs.35 %>%
  
  pivot_wider(names_from = "parcelset", values_from = "b") %>%
  group_by(contrast) %>%
  summarize(
    p = t.test(dmcc35, null34, paired = TRUE)$p.value,
    statistic = t.test(dmcc35, null34, paired = TRUE)$statistic,
    b = mean(dmcc35) - mean(null34),
    .groups = "drop_last"
    )

library(lemon)

p <-
  contrasts.subjs.35 %>%
  
  filter(contrast %in% c("hihi_lolo", "hihi_hilo")) %>%
  
  ggplot(aes(contrast, b, fill = parcelset)) +
  
  geom_hline(yintercept = 0) +
  
  stat_summary(
    fun = mean, geom = "bar", position = position_dodge(width = 1),
    width = 0.5
    ) +
  stat_summary(
    fun.data = "mean_cl_boot", 
    geom = "errorbar", position = position_dodge(width = 1), width = 0, size = 1
    ) +
  
  scale_fill_discrete_qualitative("Dark 2") +
  
  annotate(
    geom = "text", label = "DMCC35", x = 0.5, y = 0.13, fontface = "bold", size = 5, 
    hjust = 0,
    color = colorspace::qualitative_hcl(2, palette = "dark2")[1]
    ) +
  
  annotate(
    geom = "text", label = "Null34", x = 0.5, y = 0.11, fontface = "bold", size = 5, 
    hjust = 0,
    color = colorspace::qualitative_hcl(2, palette = "dark2")[2]
    ) +

  labs(y = "Mean contrast\non correlations (z)", x = "Contrast") +
  scale_x_discrete(labels = c("HiHi–HiLo", "HiHi–LoLo")) +
  
  theme_minimal(base_size = 14) +
  
  theme(
   legend.position = "none", 
   panel.grid      = element_blank(), 
   panel.border    = element_blank(),
   axis.line.y     = element_line(),
   axis.ticks.y    = element_line()
  ) +
  
  coord_capped_cart(left = "both") +
  # scale_y_continuous(breaks = c(0, 0.04, 0.08)) +
    
  annotate(geom = "segment", x = 0.75, xend = 1.25, y = -0.01, yend = -0.01, size = 0.75) +
  annotate(geom = "segment", x = 1.75, xend = 2.25, y = -0.01, yend = -0.01, size = 0.75) 

p

ggsave(
  here("code", "cross-task-simil_jocn2021", "fig_crosstasksimil.tiff"),
  p, device = "tiff", dpi = 1200, width = 8, height = 6, unit = "cm")


```


## dmcc 35 condition-wise means

```{r}

p.conditional <- contrasts.subjs %>%
  
  filter(
    parcel %in% parcellation$key[dmcc35], 
    contrast %in% c("hihi", "hilo", "lolo")
  ) %>%

  group_by(subj, contrast) %>%
  summarize(b = mean(b)) %>%
  

  ggplot(aes(contrast, b)) +
  
  geom_hline(yintercept = 0) +
  
  # geom_line(aes(group = subj), alpha = 0.25) +
  
  stat_summary(
    fun = mean, geom = "bar", fill = "grey60",
    width = 0.5
    ) +
  stat_summary(
    fun.data = "mean_cl_boot", 
    geom = "errorbar", width = 0, size = 1
    ) +
  
  labs(y = "Mean correlation (z)", x = "Condition pair") +
  scale_x_discrete(labels = c("Cor(Hi, Hi)", "Cor(Hi, Lo)", "Cor(Lo, Lo)")) +
  
  theme_minimal(base_size = 14) +
  
  theme(
   legend.position = "none", 
   panel.grid      = element_blank(), 
   panel.border    = element_blank(),
   axis.line.y     = element_line(),
   axis.ticks.y    = element_line(),
   axis.text.x = element_text(size = rel(0.75))
  ) +
  
  coord_capped_cart(left = "both") +
  
  annotate(geom = "segment", x = 1, xend = 2, y = -0.04, yend = -0.04, size = 0.75) +
  annotate(geom = "segment", x = 1, xend = 3, y = -0.06, yend = -0.06, size = 0.75) +
  annotate(geom = "segment", x = 2, xend = 3, y = -0.08, yend = -0.08, size = 0.75)

  

p.conditional

ggsave(
  here("code", "cross-task-simil_jocn2021", "fig_crosstasksimil_conditoinal.tiff"),
  p.conditional, device = "tiff", dpi = 1200, width = 8, height = 6, unit = "cm")




```






# univariate effects

```{r}


contrasts.subjs.univa <- stats.subjs.univa %>%
  
  pivot_wider(names_from = "term", values_from = "beta") %>%
  
  mutate(
    hihi_lolo = hihi - lolo,
    hihi_hilo = hihi - hilo,
    lolo_hilo = lolo - hilo
    ) %>%
  
  pivot_longer(
    cols = c("hihi_hilo", "lolo_hilo", "hihi", "lolo", "hilo", "hihi_lolo"), 
    names_to = "contrast", values_to = "b"
    ) %>%
  
  filter(parcel %in% parcellation$key[c(dmcc35, null34)], contrast %in% c("hihi_lolo", "hihi_hilo")) %>%
  mutate(
    parcelset = ifelse(parcel %in% parcellation$key[dmcc35], "dmcc35", "null34")
  ) %>%
  
  group_by(subj, contrast, parcelset) %>%
  summarize(b = mean(b))


contrasts.subjs.unifo <- stats.subjs.unifo %>%
  
  pivot_wider(names_from = "term", values_from = "beta") %>%
  
  mutate(
    hihi_lolo = hihi - lolo,
    hihi_hilo = hihi - hilo,
    lolo_hilo = lolo - hilo
    ) %>%
  
  pivot_longer(
    cols = c("hihi_hilo", "lolo_hilo", "hihi", "lolo", "hilo", "hihi_lolo"), 
    names_to = "contrast", values_to = "b"
    ) %>%
  
  filter(parcel %in% parcellation$key[c(dmcc35, null34)], contrast %in% c("hihi_lolo", "hihi_hilo")) %>%
  mutate(
    parcelset = ifelse(parcel %in% parcellation$key[dmcc35], "dmcc35", "null34")
  ) %>%
  
  group_by(subj, contrast, parcelset) %>%
  summarize(b = mean(b))



contrasts.subjs.univariate <- full_join(
  contrasts.subjs.univa %>% 
  rename(univa = b), contrasts.subjs.unifo %>% rename(unifo = b))

contrasts.subjs.35 <- full_join(contrasts.subjs.35, contrasts.subjs.univariate)


# contrasts.subjs.35 %>%
#   
#   filter(contrast == "hihi_hilo", parcelset == "dmcc35") %>%
#   ungroup %>%
#   select(b, univa, unifo) %>%
#   cor


contrasts.subjs.35 %>%
  
  filter(contrast == "hihi_hilo", parcelset == "dmcc35") %>%
  ggplot(aes(b, univa)) +
  geom_point() +
  labs(x = "multivariate contrast", y = "univariate contrast", title = "RSA contrast: hihi minus hilo")


```
